app/requirements.txt
alembic==1.10.3
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.6
pydantic==1.10.7
PyJWT==2.6.0
pytest==7.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
